[IMP] Añadido los modelos de Rol,Board,Task,Tasklist, algunos de sus repositorios y servicios. [IMP] Añadido a la página principal la carga de los tableros de usuario

+++ app/src/main/java/com/example/protasks/models/IUser.kt
@@ -11,7 +11,7 @@ interface IUser {
 
     fun getSurname(): String?
 
-    fun getId(): Int?
+    fun getId(): Long?
 
 
 }
